backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
alembic==1.14.0
paho-mqtt>=1.6,<2.0
